refactor, code rev
