+++ sdl2-sonic-drivers/CMakeLists.txt
@@ -138,7 +138,7 @@ target_sources(${LIB_NAME} PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/src/drivers/midi/adlib/OplChannel.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/drivers/midi/adlib/OplVoice.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/drivers/midi/devices/Adlib.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/drivers/midi/devices/opl/OplWriter.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/drivers/opl/OplWriter.cpp
     # --- #
     ${CMAKE_CURRENT_SOURCE_DIR}/src/files/File.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/files/IFFFile.cpp
